refactor: reorganize Latent Insights demo into modular file structure
Extract constants to config.ts, pure helpers to utils.ts, useLocalSessions
to hooks/, and split monolithic index.tsx (744 lines) into Dashboard.tsx +
LandingScreen.tsx. Remove duplicate API_BASE, centralize magic numbers and
color palettes, delete unused UploadButton. Net -1,148 lines from existing
files while preserving all functionality.
