+++ src/pages/Demos/LatentInsights/types.ts
@@ -128,21 +128,3 @@ export interface SelectedNode {
   eventIndex?: number;
   threadStatus?: string;
 }
-
-export const FEATURED_SESSIONS = [
-  {
-    id: "846f0bbfefc0",
-    dataset: "cars.csv",
-    description: "10 threads · 60 steps",
-  },
-  {
-    id: "a59dfbbd0fee",
-    dataset: "exoplanets-nasa.csv",
-    description: "8 threads · 4 waiting",
-  },
-  {
-    id: "746fa2380425",
-    dataset: "star_classification.csv",
-    description: "8 threads · 3 waiting",
-  },
-] as const;
